[clang-tidy] Modernise namespaces
Modernising all namespaces in brave to use nested concat namespace
declarations.

This change is part of a large effort to modernise the brave codebase
across the board with `clang-tidy`, in order to allow future use of
`clang-duty`during code reviews. The idea is to apply a few modernisers
across the codebase in order to have less noise when trying to use
clang-tidy with some sort of autmoation.